requirements.txt
Django==3.2.16
PyMySQL==0.9.3
reportlab
django-crispy-forms==1.14
openpyxl
cython
numpy==1.23.3
numpy-financial==1.0.0
crispy-bootstrap5
python-dateutil
django-registration-redux
django-tempus-dominus
django-crontab
django-mathfilters
django-suit==0.2.25
requests
material
django_cron
xhtml2pdf
gunicorn
pytesseract
pdf2image
python-docx
python-dotenv
twilio
openai
httpx
